Default Dirs, POST/GET Vars, Expect Script
Changed the default directories to match the file name.

All POST/GET variables are now pulled based on 'cmd'. Only the ones
required for the given command are extracted from the url.

All Expect scripts have been replaced with piped passwords to su/sudo.
This resolves the case where the password is visible in ps.

Resolves #0003, #0012, #0013